backend/requirements.txt
fastapi
uvicorn[standard]
kafka-python
pandas
python-dotenv
requests
